feat: add header ctas
