+++ sanity/schemas/components/button.ts
@@ -1,4 +1,4 @@
-import { defineField, defineType } from 'sanity'
+import { defineArrayMember, defineField, defineType } from 'sanity'
 
 import { assert } from '@/utils/assert'
 import { eitherOr } from '../utils'
@@ -6,7 +6,6 @@ import type { ImageWithAlt } from './image'
 
 export const button = defineType({
   name: 'button',
-  title: 'Button',
   type: 'object',
   fields: [
     defineField({
@@ -66,3 +65,49 @@ export type Button = {
   icon?: ImageWithAlt
   link: string
 }
+
+export const ctas = defineType({
+  name: 'ctas',
+  type: 'array',
+  validation: (rule) =>
+    rule.custom((field) => {
+      assert(Array.isArray(field))
+
+      if (field.length === 1) return true
+      if (field.length > 2) return 'No more than two buttons are allowed'
+
+      const primaryButtons = field.reduce(
+        (count: number, item: any) => count + (item.primary ? 1 : 0),
+        0,
+      )
+
+      if (primaryButtons > 1) return 'Only one primary button is allowed'
+      if (primaryButtons < 1) return 'One button must be marked as primary'
+
+      return true
+    }),
+  of: [
+    defineArrayMember({
+      name: 'cta',
+      title: 'Call to action',
+      type: 'object',
+      fields: [
+        defineField({
+          name: 'primary',
+          title: 'Primary call to action (shown on mobile)',
+          type: 'boolean',
+        }),
+        defineField({
+          name: 'button',
+          type: 'button',
+        }),
+      ],
+    }),
+  ],
+})
+
+export type Ctas = {
+  _key: string
+  primary: boolean | undefined
+  button: Button
+}[]
